Move MediaStreamConstraints into separate module
